+++ .gitignore
@@ -1,2 +1,1 @@
-cmd/cmd
-cmd/osversion
+cmd.apk
