+++ go.mod
@@ -1,0 +1,20 @@
+module github.com/getlantern/osversion
+
+go 1.18
+
+require (
+	github.com/shirou/gopsutil/v3 v3.22.12
+	golang.org/x/mobile v0.0.0-20221110043201-43a038452099
+)
+
+require (
+	github.com/go-ole/go-ole v1.2.6 // indirect
+	github.com/lufia/plan9stats v0.0.0-20211012122336-39d0f177ccd0 // indirect
+	github.com/power-devops/perfstat v0.0.0-20210106213030-5aafc221ea8c // indirect
+	github.com/tklauser/go-sysconf v0.3.11 // indirect
+	github.com/tklauser/numcpus v0.6.0 // indirect
+	github.com/yusufpapurcu/wmi v1.2.2 // indirect
+	golang.org/x/exp v0.0.0-20190731235908-ec7cb31e5a56 // indirect
+	golang.org/x/image v0.0.0-20190802002840-cff245a6509b // indirect
+	golang.org/x/sys v0.3.0 // indirect
+)
